@@ -32,10 +32,10 @@ enum BarcodeFormats {
 // ignore: must_be_immutable
 class CameraKitView extends StatefulWidget {
   /// In barcodeReader mode, while camera preview detect barcodes, This method is called.
-  final Function onBarcodeRead;
+  final Function? onBarcodeRead;
 
   ///After android and iOS user deny run time permission, this method is called.
-  final Function onPermissionDenied;
+  final Function? onPermissionDenied;
 
   ///There are 2 modes `ScaleTypeMode.fill` and `ScaleTypeMode.fit` for this parameter.
   ///If you want camera preview fill your widget area, use `fill` mode. In this mode, camera preview may be croped for filling widget area.
@@ -54,7 +54,7 @@ class CameraKitView extends StatefulWidget {
   final BarcodeFormats barcodeFormat;
 
   ///Controller for this widget
-  final CameraKitController cameraKitController;
+  final CameraKitController? cameraKitController;
 
   ///This parameter has been replaced with `useCamera2API`.
   ///This parameter accepts 3 values, `API_X`, `API_1`, `API_2`. Default value is `API_X`.
@@ -67,10 +67,10 @@ class CameraKitView extends StatefulWidget {
   ///Set front and back camera
   final CameraSelector cameraSelector;
 
-  _BarcodeScannerViewState viewState;
+  late _BarcodeScannerViewState viewState;
 
   CameraKitView(
-      {Key key,
+      {Key? key,
       this.hasBarcodeReader = false,
       this.scaleType = ScaleTypeMode.fill,
       this.onBarcodeRead,
@@ -88,7 +88,7 @@ class CameraKitView extends StatefulWidget {
 
   @override
   State<StatefulWidget> createState() {
-    if (cameraKitController != null) cameraKitController.setView(this);
+    if (cameraKitController != null) cameraKitController!.setView(this);
     viewState = _BarcodeScannerViewState();
     return viewState;
   }
@@ -96,22 +96,22 @@ class CameraKitView extends StatefulWidget {
 
 class _BarcodeScannerViewState extends State<CameraKitView>
     with WidgetsBindingObserver {
-  NativeCameraKitController controller;
-  VisibilityDetector visibilityDetector;
+  NativeCameraKitController? controller;
+  late VisibilityDetector visibilityDetector;
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     if (defaultTargetPlatform == TargetPlatform.android) {
       visibilityDetector = VisibilityDetector(
           key: Key('visible-camerakit-key-1'),
           onVisibilityChanged: (visibilityInfo) {
             if (controller != null) {
               if (visibilityInfo.visibleFraction == 0)
-                controller.setCameraVisible(false);
+                controller!.setCameraVisible(false);
               else
-                controller.setCameraVisible(true);
+                controller!.setCameraVisible(true);
             }
           },
           child: AndroidView(
@@ -123,9 +123,9 @@ class _BarcodeScannerViewState extends State<CameraKitView>
           key: Key('visible-camerakit-key-1'),
           onVisibilityChanged: (visibilityInfo) {
             if (visibilityInfo.visibleFraction == 0)
-              controller.setCameraVisible(false);
+              controller!.setCameraVisible(false);
             else
-              controller.setCameraVisible(true);
+              controller!.setCameraVisible(true);
           },
           child: UiKitView(
             viewType: 'plugins/camera_kit',
@@ -144,17 +144,17 @@ class _BarcodeScannerViewState extends State<CameraKitView>
     switch (state) {
       case AppLifecycleState.resumed:
         print("Flutter Life Cycle: resumed");
-        if (controller != null) controller.resumeCamera();
+        if (controller != null) controller!.resumeCamera();
         break;
       case AppLifecycleState.inactive:
         print("Flutter Life Cycle: inactive");
         if (Platform.isIOS) {
-          controller.pauseCamera();
+          controller!.pauseCamera();
         }
         break;
       case AppLifecycleState.paused:
         print("Flutter Life Cycle: paused");
-        controller.pauseCamera();
+        controller!.pauseCamera();
         break;
       default:
         break;
@@ -163,17 +163,17 @@ class _BarcodeScannerViewState extends State<CameraKitView>
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 
   void _onPlatformViewCreated(int id) {
     this.controller = new NativeCameraKitController._(id, context, widget);
-    this.controller.initCamera();
+    this.controller!.initCamera();
   }
 
   void disposeView() {
-    controller.dispose();
+    controller!.dispose();
   }
 }
 
@@ -191,7 +191,7 @@ class NativeCameraKitController {
   Future<dynamic> nativeMethodCallHandler(MethodCall methodCall) async {
     if (methodCall.method == "onBarcodeRead") {
       if (widget.onBarcodeRead != null)
-        widget.onBarcodeRead(methodCall.arguments);
+        widget.onBarcodeRead!(methodCall.arguments);
     }
 
     return null;
@@ -204,8 +204,8 @@ class NativeCameraKitController {
       return false;
   }
 
-  String _getCharFlashMode(CameraFlashMode cameraFlashMode) {
-    String flashMode;
+  String? _getCharFlashMode(CameraFlashMode cameraFlashMode) {
+    String? flashMode;
     switch (cameraFlashMode) {
       case CameraFlashMode.auto:
         flashMode = "A";
@@ -264,7 +264,7 @@ class NativeCameraKitController {
           });
         }
       } else {
-        widget.onPermissionDenied();
+        widget.onPermissionDenied!();
       }
     });
   }
@@ -285,7 +285,7 @@ class NativeCameraKitController {
   }
 
   ///Call take picture in Native API
-  Future<String> takePicture() async {
+  Future<String?> takePicture() async {
     return _channel.invokeMethod('takePicture', null);
   }
 
